use sig_t to declare signal handler func ptrs
If there is a platform for which 'sig_t' is an undeclared type,
then C-Kermit should declare 'sig_t' for that platform.

Declaring stack variables as

  SIGTYP (*var)();

is incorrect for a CK_ANSIC build.   Replacing all definitions with

  #ifdef CK_ANSIC
    SIGTYP (*var)(int);
  #else
    SIGTYP (*var)();
  #endif

would work but should be unnecessary when sig_t if available.  sig_t
is available anywhere ckcftp.c is built.  See discussion regarding
my_sig_t in that source file.

Change-Id: I2027fd4b4e556765d8d2c4e384be5af90f339b01