ck_ssl.c: ssl_once_init do not exclude OS2ONLY from version check
In the commercial days of Kermit 95, the OS/2 build was staticly
linked against the OpenSSL libraries.  As such, there could be no
version mismatch and hence there was no benefit to performing a
run-time check of OpenSSL ABI compatibility.

Now that Kermit 95 can be built by anyone and it can be built against
a dynamically linked set of libraries, the version check is applicable.