Create change
GitHub Pull Request: https://github.com/KermitProject/ckermit/pull/16

Fix CLEAR APC command

The `CLEAR APC` commands documented purpose is to cause Kermit to remain on the command screen after executing an APC command. As such, this command itself is allowed to be executed by APC.

However, the mechanim the command uses to achieve its goal has an unfortunate side-effect: for the rest of the APC sequence, Kermit now thinks the commands its executing came from the user at the command screen - not the remote host. As a result it stops checking whether subsequent commands are allowed to be executed via APC which is a problem. As APC is off by default the overall impact should be small.

The fix is to set a dedicated flag instead of clearing the apc status, and defer apc clearing until the very end of the APC sequence.

If APC is currently set to UNCHECKED then on non-K95 platforms the former behaviour is preserved as in this state the host can already run what it likes and I'm not sure if actually clearing the APC status mid-APC sequence may have desirable effects in some situations. K95 platforms are excluded as clearing the APC status from inside an APC results in a deadlock (this is how I stumbled on the issue).

Patch-set: 1
Change-id: I6124d2d8bf3535788d414199235e239141c1c163
Subject: This fixes a problem with CLEAR APC that results in it defeating all APC safety checks.
Branch: refs/heads/main
Status: new
Topic: 
Commit: 765bb012efa7d31d7af32135212233ba2b563d18
Tag: autogenerated:gerrit:newPatchSet
Groups: 765bb012efa7d31d7af32135212233ba2b563d18
Private: false
Work-in-progress: false